after merge brian

+++ run.py
@@ -17,6 +17,7 @@
 import sys
 from bottle import run
 from bottle import route, get, post, request, static_file, error, Bottle, template
+import bottle.ext.sqlite
 import model
 import argparse
 import bottle
@@ -46,7 +47,7 @@ except ImportError:
 
 # It might be a good idea to move the following settings to a config file and then load them
 # Change this to your IP address or 0.0.0.0 when actually hosting
-host = 'localhost' if default_configs else configs["web"]["host"]
+host = '0.0.0.0' if default_configs else configs["web"]["host"]
 
 # Test port, change to the appropriate port to host
 port = 8080 if default_configs else configs["web"]["port"]
@@ -70,11 +71,13 @@ def do_index():
 
 @get('/resource/download/<filename>')
 def do_download(filename):
+    model.aaa.require(fail_redirect='/login')
     """Return a static file from the files directory"""
     return bottle.static_file(filename, root=app.config['file_upload.dir'])
 
 @post('/resource/upload')
 def do_upload():
+    model.aaa.require(fail_redirect='/login')
     """Upload a file if it's missing"""
     upload = bottle.request.files.get('upload') # pylint: disable-msg=E1101
     try:
@@ -87,6 +90,9 @@ def do_upload():
 
 @get('/resource/delete/<filename>')
 def do_delete(filename):
+    model.aaa.require(fail_redirect='/login')
+    if model.aaa.current_user.role == "user":
+        return
     os.remove("files/" + filename)
     bottle.redirect('/resource')
 
@@ -103,6 +109,12 @@ def run_server():
         Runs a bottle server
     '''
     # app = bottle.app()
+
+    # add bottle-sqlite plugin
+    # link to sqlite3 database
+    plugin=bottle.ext.sqlite.Plugin(dbfile='./database/info2222.db')
+    app.install(plugin)
+
     session_opts = {
         'session.cookie_expires': True,
         'session.encrypt_key': 'please use a random key and keep it secret!',
@@ -114,9 +126,9 @@ def run_server():
 
 ################################################################################################################
 
-    bottle.route('/resource', 'GET', do_index)
-    bottle.route('/resource/download/<filename>', 'GET', do_download)
-    bottle.route('/resource/upload', 'POST', do_upload)
+    # bottle.route('/resource', 'GET', do_index)
+    # bottle.route('/resource/download/<filename>', 'GET', do_download)
+    # bottle.route('/resource/upload', 'POST', do_upload)
 
     # Change working directory so relative paths (and template lookup) work
     os.chdir(os.path.dirname(os.path.abspath(__file__)))
